requirements.txt
pyyaml>=6.0
requests>=2.28.0
python-dotenv>=1.0.0 
pdfkit>=0.8.0
wkhtmltopdf>=0.12.0